split up server.py
