+++ custom_components/elegoo_printer/websocket/__init__.py
@@ -1,1 +1,6 @@
-"""Elegoo Websocket code."""
+"""
+Elegoo Printer WebSocket Package.
+
+This package provides WebSocket client and server components for communicating
+with Elegoo 3D printers, including a multi-printer proxy server.
+"""
