+++ custom_components/elegoo_printer/websocket/server/__init__.py
@@ -1,0 +1,17 @@
+"""
+Elegoo Printer Proxy Server Components.
+
+This package contains all the server-side components for the multi-printer
+proxy server including the main server, printer registry, discovery protocol,
+and shared utilities.
+"""
+
+from .discovery import DiscoveryProtocol
+from .proxy import ElegooPrinterServer
+from .registry import PrinterRegistry
+
+__all__ = [
+    "DiscoveryProtocol",
+    "ElegooPrinterServer",
+    "PrinterRegistry",
+]
